Update documentation files and minor refactoring
