+++ sdgym/synthesizers/utils.py
@@ -1,4 +1,5 @@
 import subprocess
+import types
 
 import numpy as np
 import pandas as pd
@@ -7,6 +8,87 @@ from sklearn.mixture import BayesianGaussianMixture, GaussianMixture
 from sklearn.preprocessing import KBinsDiscretizer
 
 from sdgym.constants import CATEGORICAL, CONTINUOUS, ORDINAL
+from sdgym.errors import SDGymError
+from sdgym.synthesizers.base import Baseline
+
+
+def _get_synthesizer_name(synthesizer):
+    """Get the name of the synthesizer function or class.
+
+    If the given synthesizer is a function, return its name.
+    If it is a method, return the name of the class to which
+    the method belongs.
+
+    Args:
+        synthesizer (function or method):
+            The synthesizer function or method.
+
+    Returns:
+        str:
+            Name of the function or the class to which the method belongs.
+    """
+    if isinstance(synthesizer, types.MethodType):
+        synthesizer_name = synthesizer.__self__.__class__.__name__
+    else:
+        synthesizer_name = synthesizer.__name__
+
+    return synthesizer_name
+
+
+def _get_synthesizer(synthesizer, name=None):
+    if isinstance(synthesizer, str):
+        baselines = Baseline.get_subclasses()
+        if synthesizer in baselines:
+            synthesizer = baselines[synthesizer]
+        else:
+            try:
+                synthesizer = import_object(synthesizer)
+            except Exception:
+                raise SDGymError(f'Unknown synthesizer {synthesizer}') from None
+
+    if name:
+        synthesizer.name = name
+    elif not hasattr(synthesizer, 'name'):
+        synthesizer.name = _get_synthesizer_name(synthesizer)
+
+    return synthesizer
+
+
+def get_synthesizers_dict(synthesizers):
+    """Get the dict of synthesizers from the input value.
+
+    If the input is a synthesizer or an iterable of synthesizers, get their names
+    and put them on a dict.
+
+    Args:
+        synthesizers (function, class, list, tuple or dict):
+            A synthesizer (function or method or class) or an iterable of synthesizers
+            or a dict containing synthesizer names as keys and synthesizers as values.
+
+    Returns:
+        dict[str, function]:
+            dict containing synthesizer names as keys and function as values.
+
+    Raises:
+        TypeError:
+            if neither a synthesizer or an iterable or a dict is passed.
+    """
+    if callable(synthesizers):
+        return [_get_synthesizer(synthesizers)]
+
+    if isinstance(synthesizers, (list, tuple)):
+        return [
+            _get_synthesizer(synthesizer)
+            for synthesizer in synthesizers
+        ]
+
+    if isinstance(synthesizers, dict):
+        return [
+            _get_synthesizer(synthesizer, name)
+            for name, synthesizer in synthesizers.items()
+        ]
+
+    raise TypeError('`synthesizers` can only be a function, a class, a list or a dict')
 
 
 class Transformer:
